requirements.txt
flask
waitress
libsql
vercel_blob
pillow
APScheduler